Time: 5 ms (19.26%), Space: 36.4 MB (21.94%) - LeetHub

+++ 1899-count-items-matching-a-rule/1899-count-items-matching-a-rule.cpp
@@ -11,7 +11,7 @@ public:
     }
 
     int countMatches(vector<vector<string>>& items, string ruleKey, string ruleValue) {
-        unsigned short answer = 0;
+        int answer = 0;
 
         for (auto &item : items)
             if (item[um[ruleKey]] == ruleValue)
